Переход на rv64

+++ include/kernelSpace/arch/riscV/common.h
@@ -4,10 +4,10 @@
 #include <stdint.h>
 
 extern struct scause_t read_csr_scause(void);
-extern uint32_t read_csr_stval(void);
-extern uint32_t read_csr_sepc(void);
-extern uint32_t read_csr_sie(void);
+extern uint64_t read_csr_stval(void);
+extern uint64_t read_csr_sepc(void);
+extern uint64_t read_csr_sie(void);
 
-extern void write_csr_sie(uint32_t);
+extern void write_csr_sie(uint64_t);
 
 #endif
