+++ kernelSpace/arch/riscV/memory/alloc.c
@@ -1,11 +1,11 @@
 #include "../../../../include/kernelSpace/arch/riscV/memory/alloc.h"
 
-unsigned char* alloc_pages(uint32_t n) {
+unsigned char* alloc_pages(uint64_t n) {
 	if(n <= 0) PANIC("Error alloc page. n <= 0");
 
-	uint32_t sz = n;
+	uint64_t sz = n;
 	
-	int i = 0, endAddr = ((uint32_t)_free_ram_end - (uint32_t)_free_ram_start) / PAGE_SIZE;
+	int i = 0, endAddr = ((uint64_t)_free_ram_end - (uint64_t)_free_ram_start) / PAGE_SIZE;
 
 	for(; i < endAddr && sz > 0; ++i) {
 		if(!buffFreeRam[i]) --sz;
@@ -20,19 +20,19 @@ unsigned char* alloc_pages(uint32_t n) {
 	int end = i;
 	for(; start < end; ++start) buffFreeRam[start] = true;
 
-	uint32_t addr = (((uint32_t)_free_ram_start) + ((i - n) * PAGE_SIZE));
+	uint64_t addr = (((uint64_t)_free_ram_start) + ((i - n) * PAGE_SIZE));
 
 	memset((void*)addr, 0, n * PAGE_SIZE);
 
 	return (unsigned char*)addr;
 }
 
-void dealloc_page(unsigned char* phys_addr, uint32_t n) {
+void dealloc_page(unsigned char* phys_addr, uint64_t n) {
 	if(n <= 0) PANIC("Error dealloc page. n <= 0");
 
-	if((uint32_t)phys_addr < (uint32_t)_free_ram_start || (uint32_t)phys_addr >= (uint32_t)_free_ram_end) PANIC("Error dealloc page. Address invalid");
+	if((uint64_t)phys_addr < (uint64_t)_free_ram_start || (uint64_t)phys_addr >= (uint64_t)_free_ram_end) PANIC("Error dealloc page. Address invalid");
 
-	uint32_t addr = ((uint32_t)phys_addr - (uint32_t)_free_ram_start) / PAGE_SIZE;
+	uint64_t addr = ((uint64_t)phys_addr - (uint64_t)_free_ram_start) / PAGE_SIZE;
 
 	if(addr + n > MAX_RAM) PANIC("Error dealloc page. Address is full");
